+++ Modules/Microsoft365DSC/DscResources/MSFT_EXOIRMConfiguration/MSFT_EXOIRMConfiguration.psm1
@@ -116,6 +116,12 @@ function Get-TargetResource
 
     Write-Verbose -Message 'Getting IRM Configuration'
 
+    if ($PSBoundParameters.ContainsKey('EnablePortalTrackingLogs'))
+    {
+        $PSBoundParameters.Remove('EnablePortalTrackingLogs') | Out-Null
+        Write-Warning "Property 'EnablePortalTrackingLogs' is deprecated and will be removed“
+    }
+
     try
     {
         if (-not $Script:exportedInstance)
@@ -325,6 +331,12 @@ function Set-TargetResource
 
     Write-Verbose -Message 'Setting configuration of Resource Configuration'
 
+    if ($PSBoundParameters.ContainsKey('EnablePortalTrackingLogs'))
+    {
+        $PSBoundParameters.Remove('EnablePortalTrackingLogs') | Out-Null
+        Write-Warning "Property 'EnablePortalTrackingLogs' is deprecated and will be removed“
+    }
+
     $null = New-M365DSCConnection -Workload 'ExchangeOnline' `
         -InboundParameters $PSBoundParameters
 
@@ -463,8 +475,10 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
+    $compareParameters = Get-CompareParameters
     $result = Test-M365DSCTargetResource -DesiredValues $PSBoundParameters `
-        -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '')
+        -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '') `
+        @compareParameters
     return $result
 }
 
@@ -569,4 +583,16 @@ function Export-TargetResource
         throw
     }
 }
-Export-ModuleMember -Function *-TargetResource
+
+function Get-CompareParameters
+{
+    [CmdletBinding()]
+    [OutputType([System.Collections.Hashtable])]
+    param()
+
+    return @{
+        ExcludedProperties = @('EnablePortalTrackingLogs')
+    }
+}
+
+Export-ModuleMember -Function @('*-TargetResource', 'Get-CompareParameters')